refactor(user-dashboard): break up files
