+++ packages/user-dashboard/src/components/consent/hooks/useVersionEnabledCheck.ts
@@ -6,29 +6,34 @@ import { useUrlAppId } from './useUrlAppId';
  * Hook to check whether an app version is enabled.
  * Automatically performs the check on mount.
  */
-export const useVersionEnabledCheck = ({ 
-  versionNumber, 
-  specificAppId 
-}: { 
+export const useVersionEnabledCheck = ({
+  versionNumber,
+  specificAppId,
+}: {
   versionNumber: number;
-  specificAppId?: string | number; 
+  specificAppId?: string | number;
 }) => {
   const { appId: urlAppId } = useUrlAppId();
   const appId = specificAppId || urlAppId;
   const [isVersionEnabled, setIsVersionEnabled] = useState<boolean | null>(null);
 
   useEffect(() => {
+    // Skip invalid inputs
+    if (!appId || !versionNumber || versionNumber === 0) {
+      setIsVersionEnabled(null);
+      return;
+    }
+
     const checkVersionEnabled = async () => {
-      if (!appId || Number(appId) === 0) {
-        return;
-      }
-      
       try {
         const contract = getAppViewRegistryContract();
         const [, versionData] = await contract.getAppVersion(Number(appId), versionNumber);
         setIsVersionEnabled(versionData.enabled);
-      } catch (err) {
-        console.error('Error checking if version is enabled:', err);
+      } catch (error: any) {
+        // Handle AppVersionNotRegistered quietly
+        if (error?.errorName !== 'AppVersionNotRegistered') {
+          console.warn(`Version check error for app ${appId}, version ${versionNumber}`);
+        }
         setIsVersionEnabled(null);
       }
     };
@@ -37,4 +42,4 @@ export const useVersionEnabledCheck = ({
   }, [appId, versionNumber]);
 
   return { isVersionEnabled };
-}; 
+};
